fix data classes create data classes for all dataBase entity
