+++ app/src/main/java/ru/gb/bufet/model/responseData/AdvertiseBanners.kt
@@ -1,0 +1,9 @@
+package ru.gb.bufet.model.responseData
+
+data class AdvertiseBanners(
+    val description: String?,
+    val id: Int?,
+    val image: String?,
+    val isActive: Boolean?,
+    val title: String?
+)
