+++ app/src/main/java/ru/gb/bufet/model/responseData/RestaurantPicture.kt
@@ -1,0 +1,7 @@
+package ru.gb.bufet.model.responseData
+
+data class RestaurantPicture(
+    val id: Int?,
+    val imagePath: String?,
+    val restaurantId: Int?
+)
