+++ app/src/main/java/ru/gb/bufet/model/responseData/RestaurantTable.kt
@@ -1,0 +1,9 @@
+package ru.gb.bufet.model.responseData
+
+data class RestaurantTable(
+    val availability: Int?,
+    val guestId: Int?,
+    val guestsCount: Int?,
+    val id: Int?,
+    val restaurantId: Int?
+)
